Feature/point lights (#98)
* Added PointLight logic to the engine.

Still need to add the pointlights in the shader.

* Update gpu_scene.hpp

* Added Point Light functionality to the Engine

Works with both adding component at runtime, or creating a point light in the engine.cpp itself.

Attenuation should probably be done correctly here, but there is a lot of contradicting stuff on the internet here

* Fixed ECS module change

* Changed some code based on feedback.

Still need to create a function in the shader for re-usability

* Cleaned up Shader, and now use WorldPosition

The Shader now has more functions, increasing readability.